Crate dml

This module contains the models representing the Datamodel part of a Prisma schema. It contains the main data structures that the engines can build upon.

Modules

composite_type

Composite types defined with the type keyword.

datamodel

default_value

enum

field

A field in a model.

model

native_type_constructor

native_type_instance

relation_info

scalars

traits