Enum dml::composite_type::CompositeTypeFieldType

pub enum CompositeTypeFieldType {
    CompositeType(String),
    Scalar(ScalarType, Option<NativeTypeInstance>),
    Enum(EnumId),
    Unsupported(String),
}

Variants

CompositeType(String)

Scalar(ScalarType, Option<NativeTypeInstance>)

Enum(EnumId)

This is an enum field, with an enum of the given name.

Unsupported(String)

This is a field with an unsupported datatype. The content is the db’s description of the type, it should enable migrate to create the type.

Implementations

impl CompositeTypeFieldType

pub fn as_scalar(&self) -> Option<(&ScalarType, &Option<NativeTypeInstance>)>