Struct dml::field::ScalarField

pub struct ScalarField {
    pub id: ScalarFieldId,
    pub name: String,
    pub field_type: FieldType,
    pub arity: FieldArity,
    pub database_name: Option<String>,
    pub default_value: Option<DefaultValue>,
    pub documentation: Option<String>,
    pub is_generated: bool,
    pub is_updated_at: bool,
    pub is_ignored: bool,
}

Represents a scalar field in a model.

Fields
id: ScalarFieldIdname: String

Name of the field.

field_type: FieldType

The field’s type.

arity: FieldArity

The field’s arity.

database_name: Option<String>

The database internal name.

default_value: Option<DefaultValue>

The default value.

documentation: Option<String>

Comments associated with this field.

is_generated: bool

signals that this field was internally generated (only back relation fields as of now)

is_updated_at: bool

If set, signals that this field is updated_at and will be updated to now() automatically.

is_ignored: bool

Indicates if this field is ignored by the Client.

Implementations

impl ScalarField

pub fn new(
id: ScalarFieldId,
name: &str,
arity: FieldArity,
field_type: FieldType
) -> ScalarField

Creates a new field with the given name and type.

pub fn set_default_value(&mut self, val: DefaultValue)

pub fn db_name(&self) -> &str

pub fn is_required(&self) -> bool

pub fn is_optional(&self) -> bool

pub fn is_auto_increment(&self) -> bool

pub fn default_value(&self) -> Option<&DefaultValue>