Function dml::process_block

pub fn process_block<W: Write>(txt_block: &str, output: &mut W, config: Config) -> Result<(), Error>

Process a single block of text

Examples

use dml::Config;
use dml::process_block;

let input = "# Hello World";
let mut output = std::io::stdout();
let config = Config::default();
process_block(&input, &mut output, config);

Panics

In debug mode, a empty string input, or a string input with only space cause a debug_assert panic.

Failure

Return a failure if a IO read or write as failed