Module dml::model

Structs

IndexDefinition

Represents an index defined via @@index, @unique or @@unique.

IndexField

A field in an index that optionally defines a sort order and length limit.

Model

Represents a model in a prisma schema.

PrimaryKeyDefinition

Represents a primary key defined via @@id or @id.

PrimaryKeyField

A field in a Primary Key that optionally defines a sort order and length limit.

UniqueCriteria

A unique criteria is a set of fields through which a record can be uniquely identified.

Enums

IndexAlgorithm

IndexType

OperatorClass

SortOrder