Struct dml::Datamodel

pub struct Datamodel {
    pub enums: Vec<Enum>,
    pub models: Vec<Model>,
    pub composite_types: Vec<CompositeType>,
}

Entities in the datamodel can be flagged as is_commented_out. This lets the renderer know that introspection encountered unsupported names or features and these are supposed to be rendered as comments. Since the parser will not set these flags when reading a schema string, only introspection and the lowering of the datamodel to the ast care about these flags. The FieldType: Unsupported behaves in the same way. Both of these are never converted into the internal datamodel.

Fields
enums: Vec<Enum>models: Vec<Model>composite_types: Vec<CompositeType>
Implementations

impl Datamodel

pub fn new() -> Datamodel

pub fn is_empty(&self) -> bool

Checks if a datamodel contains neither enums nor models.

pub fn add_enum(&mut self, en: Enum)

Adds an enum to this datamodel.

pub fn add_model(&mut self, model: Model)

Adds a model to this datamodel.

pub fn models(&self) -> Iter<'_, Model>

Gets an iterator over all models.

pub fn composite_types(&self) -> Iter<'_, CompositeType>

Gets an iterator over all composite types.

pub fn enums(&self) -> Iter<'_, Enum>

Gets an iterator over all enums.

pub fn models_mut(&mut self) -> IterMut<'_, Model>

Gets a mutable iterator over all models.

pub fn enums_mut(&mut self) -> IterMut<'_, Enum>

Gets a mutable iterator over all enums.

pub fn find_model(&self, name: &str) -> Option<&Model>

Finds a model by name.

pub fn find_composite_type(&self, name: &str) -> Option<&CompositeType>

Finds a composite type by name.

pub fn find_model_db_name(&self, db_name: &str) -> Option<&Model>

Finds a model by database name. This will only find models with a name remapped to the provided db_name.

pub fn find_model_by_relation_field_ref(
&self,
field: &RelationField
) -> Option<&Model>

Finds parent model for a field reference.

pub fn find_field_mut(&mut self, model: &str, field: &str) -> &mut Field

Finds a mutable field reference by a model and field name.

pub fn find_scalar_field_mut(
&mut self,
model: &str,
field: &str
) -> &mut ScalarField

Finds a mutable scalar field reference by a model and field name.

pub fn find_relation_field_mut(
&mut self,
model: &str,
field: &str
) -> &mut RelationField

Finds a mutable relation field reference by a model and field name.

pub fn find_enum(&self, name: &str) -> Option<&Enum>

Finds an enum by name.

pub fn find_enum_db_name(&self, db_name: &str) -> Option<&Enum>

Finds an enum by database name.

pub fn find_model_mut(&mut self, name: &str) -> &mut Model

Finds a model by name and returns a mutable reference.

pub fn find_enum_mut(&mut self, name: &str) -> &mut Enum

Finds an enum by name and returns a mutable reference.

pub fn find_enum_fields(&self, enum_name: &str) -> Vec<(String, String)>

Returns (model_name, field_name) for all fields using a specific enum.

pub fn find_relation_fields_for_model(
&mut self,
model_name: &str
) -> Vec<(String, String)>

Returns (model_name, field_name) for all relation fields pointing to a specific model.

pub fn find_related_field_for_info(
&self,
info: &RelationInfo,
exclude: &str
) -> Option<(usize, &RelationField)>

Finds a relation field related to a relation info. Returns a tuple (index_of_relation_field_in_model, relation_field).

pub fn find_related_field(
&self,
rf: &RelationField
) -> Option<(usize, &RelationField)>

This finds the related field for a relationfield if available

pub fn find_related_field_bang(
&self,
rf: &RelationField
) -> (usize, &RelationField)

This is used once we assume the datamodel to be internally valid