Crate dml

This module contains the models representing the Datamodel part of a Prisma schema. It contains the main data structures that the engines can build upon.

Re-exports

pub use prisma_value;

pub use self::composite_type::*;

pub use self::default_value::*;

pub use self::field::*;

pub use self::model::*;

pub use self::native_type_instance::*;

pub use self::enum::*;

pub use self::relation_info::*;

pub use self::scalars::*;

pub use self::traits::*;

Modules

composite_type

Composite types defined with the type keyword.

default_value

enum

field

A field in a model.

model

native_type_instance

relation_info

scalars

traits

Structs

Datamodel

Entities in the datamodel can be flagged as is_commented_out. This lets the renderer know that introspection encountered unsupported names or features and these are supposed to be rendered as comments. Since the parser will not set these flags when reading a schema string, only introspection and the lowering of the datamodel to the ast care about these flags. The FieldType: Unsupported behaves in the same way. Both of these are never converted into the internal datamodel.

Enums

PrismaValue

Functions

find_model_by_db_name

Find the model mapping to the passed in database name.