Crate dml

This module contains the models representing the Datamodel part of a Prisma schema. It contains the main data structures that the engines can build upon.

Re-exports

pub use prisma_value;

pub use self::composite_type::*;

pub use self::default_value::*;

pub use self::field::*;

pub use self::model::*;

pub use self::native_type_instance::*;

pub use self::relation_info::*;

pub use self::scalars::*;

pub use self::traits::*;

Modules

composite_type

Composite types defined with the type keyword.

default_value

field

A field in a model.

model

native_type_instance

relation_info

scalars

traits

Structs

Datamodel

Enums

FieldArity

An arity of a data model field.

IndexType

The different types of indexes supported in the Prisma Schema Language.

PrismaValue

ReferentialAction

An action describing the way referential integrity is managed in the system.

Functions

dml_default_kind

find_model_by_db_name

Find the model mapping to the passed in database name.

lift

Validated schema -> dml::Datamodel.